Implemented CustomList. List inside Bouquet was changed to custom list. Wrote tests for each class and hopefully most of the methods.

+++ app/src/main/java/sd/java/lab7_8/flowers/FreshFlower.java
@@ -3,7 +3,7 @@ package sd.java.lab7_8.flowers;
 public class FreshFlower extends Flower {
     private static final int FRESHNESS = 100; // Constant freshness value for FreshFlower
 
-    public FreshFlower(String color, double length) {
+    public FreshFlower(String color, int length) {
         super(color, length, FRESHNESS);
     }
 }
