+++ app/src/main/java/sd/java/lab7_8/flowers/NotSoFreshFlower.java
@@ -3,7 +3,7 @@ package sd.java.lab7_8.flowers;
 public class NotSoFreshFlower extends Flower {
     private static final int FRESHNESS = 30; // Constant freshness value for NotSoFreshFlower
 
-    public NotSoFreshFlower(String color, double length) {
+    public NotSoFreshFlower(String color, int length) {
         super(color, length, FRESHNESS);
     }
 }
